+++ tsup.config.ts
@@ -4,6 +4,8 @@ export default defineConfig({
   entry: {
     index: "src/index.ts",
     next: "src/next.ts",
+    validator: "src/validator.ts",
+    cli: "src/cli.ts",
   },
   format: ["esm", "cjs"],
   dts: {